change the content displayed in index.vue
