--- client/pages/top.vue
@@ -1,5 +1,0 @@
-<template>
-  <div>
-    <h1>top</h1>
-  </div>
-</template>
